Remove aggressive size optimization in debug builds
Enabling "codegen-units = 1" without LTO gives a duped symbol error
for "rust_begin_unwind" (generated by #[panic_handler]), implying that
the two panic handlers are aliasing in at least some configurations.
This does appear to be the case on initial inspection of the disassembly.
So far it is choosing the version at the lower address, i.e. the one in
TCM, which is fine until we run dual-core. But it's still not ideal and
should be solved.

Reverts 79eab92fed27dd39f5eeb26c9ed6b2d3ca83f540